requirements.txt
python-dotenv
pandas
numpy
scikit-learn
matplotlib
seaborn
certifi
pymongo[srv]==3.11
pymongo
# -e .
